requirements.txt
py-cord==2.6.1
python-dotenv==1.0.0
aiofiles==23.2.1
pyyaml==6.0.1
flask==3.1.1
flask-socketio==5.5.1
eventlet==0.40.1
pillow==11.3.0
requests==2.32.4
googletrans==4.0.0rc1
